Added ability to change the cloud api url
To utilize the LocalWS engine the ability to change the cloud api
url is required. LocalWS engine's API is a mirror of the SCORM
Cloud API. The initializer for ScormCloud::ScormCloud has been
updated to accept a third parameter. If passed in the URL used
for API calls will be the value of the passed in parameter.
Otherwise the default is used, this prevents any existing code
from breaking.

You can also pass in the API url from a Rails initializer,
show below.

config.scorm_cloud.api_url