Add logging functionality for request/response
Default behavior is to not log anything.

A fourth argument has been added to the initialization method that
expects a Logger (or any object that quacks like a Logger).

Alternatively you can instantiate the scorm object and then set
`logger`.

If set, the HTTP request and response will be written to the log.